+++ paddle/fluid/operators/math/CMakeLists.txt
@@ -66,9 +66,9 @@ math_library(maxouting)
 math_library(pooling)
 
 if(WITH_MKLDNN)
-    math_library(selected_rows_functor DEPS selected_rows math_function blas mkldnn_axpy_handler)
+    math_library(selected_rows_functor DEPS selected_rows_utils math_function blas mkldnn_axpy_handler)
 else()
-    math_library(selected_rows_functor DEPS selected_rows math_function blas)
+    math_library(selected_rows_functor DEPS selected_rows_utils math_function blas)
 endif()
 
 math_library(sequence2batch)
