+++ paddle/fluid/pybind/io.cc
@@ -14,7 +14,7 @@ limitations under the License. */
 
 #include "paddle/fluid/pybind/io.h"
 #include "paddle/fluid/framework/lod_tensor.h"
-#include "paddle/fluid/framework/selected_rows.h"
+#include "paddle/fluid/framework/selected_rows_utils.h"
 #include "paddle/fluid/platform/enforce.h"
 #include "paddle/fluid/pybind/pybind_boost_headers.h"
