+++ paddle/pten/api/lib/utils/CMakeLists.txt
@@ -1,2 +1,2 @@
 cc_library(pten_api_utils SRCS storage.cc tensor_utils.cc DEPS
-tensor_base convert_utils dense_tensor lod_tensor selected_rows place var_type_traits)
+tensor_base convert_utils dense_tensor lod_tensor selected_rows_utils place var_type_traits)
